+++ src/DrumPlayer.cpp
@@ -139,7 +139,7 @@ struct DrumPlayer : Module {
 		playBuffer[3][1].resize(0); 
 	}
 
-	void onReset() override {
+	void onReset(const ResetEvent &e) override {
 		interpolationMode = HERMITE_INTERP;
 		antiAlias = 1;
 		outsMode = NORMALLED_OUTS;
@@ -149,6 +149,7 @@ struct DrumPlayer : Module {
 			choking[i] = false;
 			fading[i] = false;
 		}
+		Module::onReset(e);
 	}
 
 	void onSampleRateChange() override {