+++ src/DrumPlayerXtra.cpp
@@ -215,7 +215,7 @@ struct DrumPlayerXtra : Module {
 		playBuffer[3][1].resize(0);
 	}
 
-	void onReset() override {
+	void onReset(const ResetEvent &e) override {
 		interpolationMode = HERMITE_INTERP;
 		antiAlias = 1;
 		outsMode = NORMALLED_OUTS;
@@ -244,6 +244,7 @@ struct DrumPlayerXtra : Module {
 		colorBoxB[1] = 0;
 		colorBoxB[2] = 0;
 		colorBoxB[3] = 0;
+		Module::onReset(e);
 	}
 
 	void onSampleRateChange() override {