+++ src/plugin.hpp
@@ -38,3 +38,4 @@ extern Model* modelCVrouter;
 extern Model* modelAdder8;
 extern Model* modelClocker;
 extern Model* modelSickoSampler2;
+extern Model* modelSickoAmp;